requirements.txt
anthropic
Jinja2
python-dotenv
fastapi
uvicorn
pydantic
prometheus-fastapi-instrumentator
huggingface_hub
slowapi
tenacity